+++ src/main/java/org/forweb/commandos/entity/ammo/KnifeAmmo.java
@@ -4,7 +4,7 @@ import org.forweb.commandos.controller.PersonWebSocketEndpoint;
 
 public class KnifeAmmo extends Projectile {
 
-    public KnifeAmmo(int xStart, int yStart, float angle) {
+    public KnifeAmmo(int xStart, int yStart, double angle) {
         super(xStart, yStart, angle);
         this.setLifeTime(150L);
         this.setRadius(PersonWebSocketEndpoint.PERSON_RADIUS + 15);
