+++ src/main/java/org/forweb/commandos/entity/ammo/LinkedAmmo.java
@@ -4,7 +4,7 @@ package org.forweb.commandos.entity.ammo;
 public abstract class LinkedAmmo extends Projectile{
     private int personId;
 
-    public LinkedAmmo(int xStart, int yStart, float angle, int personId) {
+    public LinkedAmmo(int xStart, int yStart, double angle, int personId) {
         super(xStart, yStart, angle);
         this.setPersonId(personId);
     }
