+++ src/main/java/org/forweb/commandos/entity/ammo/Rocket.java
@@ -3,7 +3,7 @@ package org.forweb.commandos.entity.ammo;
 public class Rocket extends LinkedAmmo {
 
 
-    public Rocket(int xStart, int yStart, float angle, int personId) {
+    public Rocket(int xStart, int yStart, double angle, int personId) {
         super(xStart, yStart, angle, personId);
         this.setLifeTime(1400L);
         this.setRadius(700);
