+++ src/main/java/org/forweb/commandos/entity/weapon/AssaultRifle.java
@@ -23,7 +23,7 @@ public class AssaultRifle extends AbstractWeapon<Bullet> {
     }
 
     @Override
-    public Projectile getProjectile(Person person, float changedAngle) {
+    public Projectile getProjectile(Person person, double changedAngle) {
         return new Bullet((int)person.getX(), (int)person.getY(), changedAngle);
     }
 }
