+++ src/main/java/org/forweb/commandos/entity/weapon/Knife.java
@@ -27,7 +27,7 @@ public class Knife extends AbstractWeapon<KnifeAmmo>{
     }
 
     @Override
-    public Projectile getProjectile(Person person, float angle) {
+    public Projectile getProjectile(Person person, double angle) {
         return new KnifeAmmo((int)person.getX(), (int)person.getY(), angle);
     }
