+++ src/main/java/org/forweb/commandos/entity/weapon/Pistol.java
@@ -22,7 +22,7 @@ public class Pistol extends AbstractWeapon<Bullet> {
     }
 
     @Override
-    public Projectile getProjectile(Person person, float angle) {
+    public Projectile getProjectile(Person person, double angle) {
         return new Bullet((int)person.getX(), (int)person.getY(), angle);
     }
 }
