+++ src/main/java/org/forweb/commandos/entity/weapon/RocketLauncher.java
@@ -22,7 +22,7 @@ public class RocketLauncher extends AbstractWeapon<Rocket> {
     }
 
     @Override
-    public Projectile getProjectile(Person person, float changedAngle) {
+    public Projectile getProjectile(Person person, double changedAngle) {
         return new Rocket((int)person.getX(), (int)person.getY(), changedAngle, person.getId());
     }
 }
