+++ src/main/java/org/forweb/commandos/entity/weapon/SniperRifle.java
@@ -22,7 +22,7 @@ public class SniperRifle extends AbstractWeapon<SniperBullet> {
     }
 
     @Override
-    public Projectile getProjectile(Person person, float changedAngle) {
+    public Projectile getProjectile(Person person, double changedAngle) {
         return new SniperBullet((int)person.getX(), (int)person.getY(), changedAngle);
     }
 }
